+++ generator/geometry_holder.hpp
@@ -188,7 +188,7 @@ private:
 
     // points conversion
     tesselator::PointsInfo points;
-    m2::PointU (*D2U)(m2::PointD const &, uint32_t) = &PointDToPointU;
+    m2::PointU (*D2U)(m2::PointD const &, uint8_t) = &PointDToPointU;
     info.GetPointsInfo(saver.GetBasePoint(), saver.GetMaxPoint(),
                        std::bind(D2U, std::placeholders::_1, cp.GetCoordBits()), points);
